move blockchain model to kernel
