+++ domain/src/main/java/onlydust/com/marketplace/api/domain/view/ReceiptView.java
@@ -2,7 +2,7 @@ package onlydust.com.marketplace.api.domain.view;
 
 import lombok.Builder;
 import lombok.Data;
-import onlydust.com.marketplace.api.domain.model.blockchain.*;
+import onlydust.com.marketplace.kernel.model.blockchain.*;
 
 import java.net.URI;
 import java.util.Optional;
