+++ infrastructure/postgres-adapter/src/test/java/onlydust/com/marketplace/api/postgres/adapter/it/repository/AllRepositoriesIT.java
@@ -4,10 +4,10 @@ import com.vladmihalcea.hibernate.type.json.internal.JacksonUtil;
 import onlydust.com.marketplace.api.domain.model.UserPayoutInformation;
 import onlydust.com.marketplace.api.domain.model.UserRole;
 import onlydust.com.marketplace.api.domain.model.bank.AccountNumber;
-import onlydust.com.marketplace.api.domain.model.blockchain.Aptos;
-import onlydust.com.marketplace.api.domain.model.blockchain.Ethereum;
-import onlydust.com.marketplace.api.domain.model.blockchain.Optimism;
-import onlydust.com.marketplace.api.domain.model.blockchain.StarkNet;
+import onlydust.com.marketplace.kernel.model.blockchain.Aptos;
+import onlydust.com.marketplace.kernel.model.blockchain.Ethereum;
+import onlydust.com.marketplace.kernel.model.blockchain.Optimism;
+import onlydust.com.marketplace.kernel.model.blockchain.StarkNet;
 import onlydust.com.marketplace.api.postgres.adapter.PostgresUserAdapter;
 import onlydust.com.marketplace.api.postgres.adapter.entity.read.UserViewEntity;
 import onlydust.com.marketplace.api.postgres.adapter.entity.write.UserEntity;
