+++ kernel/src/main/java/onlydust/com/marketplace/kernel/model/blockchain/Blockchain.java
@@ -1,4 +1,4 @@
-package onlydust.com.marketplace.api.domain.model.blockchain;
+package onlydust.com.marketplace.kernel.model.blockchain;
 
 public enum Blockchain {
     ETHEREUM, OPTIMISM, STARKNET, APTOS;
