+++ kernel/src/main/java/onlydust/com/marketplace/kernel/model/blockchain/StarkNet.java
@@ -1,6 +1,8 @@
-package onlydust.com.marketplace.api.domain.model.blockchain;
+package onlydust.com.marketplace.kernel.model.blockchain;
 
-import onlydust.com.marketplace.api.domain.model.blockchain.starknet.*;
+import onlydust.com.marketplace.kernel.model.blockchain.starknet.AccountAddress;
+import onlydust.com.marketplace.kernel.model.blockchain.starknet.StarkScan;
+import onlydust.com.marketplace.kernel.model.blockchain.starknet.TransactionHash;
 
 public interface StarkNet {
     BlockExplorer<TransactionHash> BLOCK_EXPLORER = new StarkScan();
