+++ kernel/src/main/java/onlydust/com/marketplace/kernel/model/blockchain/evm/ContractAddress.java
@@ -1,0 +1,15 @@
+package onlydust.com.marketplace.kernel.model.blockchain.evm;
+
+import lombok.EqualsAndHashCode;
+import lombok.NonNull;
+import onlydust.com.marketplace.kernel.model.blockchain.Hash;
+
+@EqualsAndHashCode(callSuper = true)
+public class ContractAddress extends Hash {
+    private static final int MAX_BYTE_COUNT = 20;
+
+    public ContractAddress(final @NonNull String address) {
+        super(MAX_BYTE_COUNT, address);
+    }
+}
+
