+++ kernel/src/main/java/onlydust/com/marketplace/kernel/model/blockchain/evm/TransactionHash.java
@@ -1,10 +1,8 @@
-package onlydust.com.marketplace.api.domain.model.blockchain.evm;
+package onlydust.com.marketplace.kernel.model.blockchain.evm;
 
 import lombok.EqualsAndHashCode;
 import lombok.NonNull;
-import onlydust.com.marketplace.api.domain.model.blockchain.Hash;
-
-import java.net.URI;
+import onlydust.com.marketplace.kernel.model.blockchain.Hash;
 
 @EqualsAndHashCode(callSuper = true)
 public class TransactionHash extends Hash {
